+++ scripts/run_param_pyuvsim.py
@@ -15,6 +15,7 @@ parser = argparse.ArgumentParser(
 )
 parser.add_argument('paramsfile', type=str, help='Parameter yaml file.', default=None)
 parser.add_argument('--profile', type=str, help='Time profiling output file name.')
+parser.add_argument('--quiet', action='store_true', help='Suppress stdout printing.')
 parser.add_argument('--raw_profile', help='Also save pickled LineStats data for line profiling.',
                     action='store_true')
 
@@ -31,7 +32,7 @@ if not os.path.isdir(os.path.dirname(args.paramsfile)):
 
 t0 = pytime.time()
 
-pyuvsim.uvsim.run_uvsim(args.paramsfile)
+pyuvsim.uvsim.run_uvsim(args.paramsfile, quiet=args.quiet)
 
 if args.profile:
     dt = pytime.time() - t0